Sprint 7 stage_8 - fix and improvement
