+++ src/main/java/order/MethodsOrder.java
@@ -1,41 +1,40 @@
 package order;
 
+import courier.BaseClient;
 import io.qameta.allure.Step;
 import io.restassured.response.ValidatableResponse;
 
 import static constants.Endpoints.*;
 import static io.restassured.RestAssured.given;
 
-public class MethodsOrder {
+public class MethodsOrder extends BaseClient {
 
     @Step("Создание заказа")
     public ValidatableResponse successOrderCreate(OrderInfo orderInfo) {
-        ValidatableResponse responseOrder = given()
+        return given()
                 .header("Content-type", "application/json")
                 .when()
                 .post(POST_ORDER_CREATE)
                 .then();
-        return responseOrder;
+
     }
 
     @Step("Отмена заказа")
     public ValidatableResponse cancelOrder(int track) {
-        ValidatableResponse responseOrder = given()
+        return given()
                 .header("Content-type", "application/json")
                 .body(track)
                 .when()
                 .put(CANCEL_ORDER)
                 .then();
-        return responseOrder;
     }
 
     @Step("Получение списка заказов")
     public ValidatableResponse getOrdersList() {
-        ValidatableResponse responseOrder = given()
+        return given()
                 .header("Content-type", "application/json")
                 .when()
                 .get(GET_ORDERS_LIST)
                 .then();
-        return responseOrder;
     }
 }
