+++ src/test/java/CreateCourierTest.java
@@ -2,22 +2,16 @@ import courier.CourierInfo;
 import courier.LoginInfo;
 import courier.MethodsCourier;
 import io.qameta.allure.Description;
-import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
-import io.restassured.RestAssured;
-import io.restassured.filter.log.RequestLoggingFilter;
-import io.restassured.filter.log.ResponseLoggingFilter;
 
 import io.restassured.response.ValidatableResponse;
-import org.junit.After;
-import org.junit.Before;
+
 import org.junit.Test;
 
-import static constants.Endpoints.*;
 import static constants.TextMessage.*;
 import static org.hamcrest.CoreMatchers.equalTo;
 
-public class CreateCourierTest {
+public class CreateCourierTest extends BasaTest {
 
     private final int random = 1 + (int) (Math.random() * 10000);
 
@@ -25,27 +19,6 @@ public class CreateCourierTest {
 
     protected MethodsCourier methodsCourier = new MethodsCourier();
     private int courierId;
-
-    @Before
-    @Step("Базовые тестовых данные")
-    public void setUp() {
-        RestAssured.filters(new RequestLoggingFilter(), new ResponseLoggingFilter());
-        RestAssured.baseURI = BASE_URL;
-    }
-
-    @After
-    @Step("Удаление ранее созданного курьера")
-    public void deleteData() {
-        if (courierId == 0) {
-            System.out.println(COURIER_EMPTY);
-        } else {
-            ValidatableResponse responseDelete = methodsCourier.courierDelete(courierId);
-            responseDelete.assertThat().statusCode(200)
-                    .body("ok", equalTo(COURIER_DELETE_200));
-            System.out.println(MESSAGE_DELETE_COURIER);
-        }
-    }
-
     @DisplayName("Создание курьера с валидными данными")
     @Description("Успешного создание курьера")
     @Test
